Derives "Therapieplan" from "Tumorkonferenz" to increase compatibility
From a cancer register perspective, a regular tumor conference is the
same as a molecular tumor board. Hence the "Therapieplan" profile from
the MTB module is derived from the Oncology module's "Tumorkonferenz"
profile, which is based on the german Onkologischer Basisdatensatz used
to report cases to the cancer register.

Only the min cardinality of two properties are affected by this change,
which should not affect overall compatibility.